Finish \tag
